Install bundler and improve the basic gem structure.
