+++ Gemfile
@@ -1,0 +1,4 @@
+source 'http://rubygems.org'
+
+# Specify your gem's dependencies in ruby_danfe.gemspec
+gemspec
