+++ README.md
@@ -1,27 +1,83 @@
-Ruby DANFE
-==========
+# Ruby DANFE
 
-This is a component for generating PDF files for Brazilian DANFE ("Documento
-Auxiliar da Nota Fiscal Eletrônica") from a valid NFE XML. 
+It generates PDF files for Brazilian DANFE (_Documento Auxiliar da Nota Fiscal Eletrônica_) from a valid NF-e XML.
 
-Inspired on NFePHP class [http://www.assembla.com/wiki/show/nfephp/DanfeNFePHP].
+It also generates Brazilian DACTE (_Documento Auxiliar do Conhecimento de Transporte Eletrônico_).
 
-Installing
-==========
+This project is inspired on [NFePHP class](http://www.assembla.com/wiki/show/nfephp/DanfeNFePHP).
 
-    gem install ruby_danfe
-    
-Usage
-=====
+## Installing
 
-    require 'ruby_danfe'
-    
-    RubyDanfe.generate('sample.pdf', 'sample.xml')
+        gem install ruby_danfe
 
-Testing
-=======
+## Usage
 
-    cd test
-    ruby generate.rb
-    ruby generate.rb nfe_with_ns.xml
-    
+If you have the xml saved in a file:
+
+        require "ruby_danfe"
+        RubyDanfe.generate("sample.pdf", "sample.xml")
+
+If you have the xml in a variable:
+
+        xml = RubyDanfe::XML.new(my_xml_string)
+        pdf = RubyDanfe.generatePDF(xml)
+        pdf.render_file "output.pdf"
+
+## Development
+
+### Installing dependencies
+
+You can install all necessaries dependencies using bunder like above:
+
+        $ bundle install
+
+### Tests
+
+At `test` folder you will find the `generate.rb` file. It shows how to generate a pdf file from a valid xml.
+
+You can use it following the steps above:
+
+        $ cd test
+        $ ruby generate.rb nfe_with_ns.xml
+
+You can also use an special version of irb with all classes pre-loaded. Just use:
+
+        $ rake console
+        RubyDanfe.generate("output.pdf", "test/nfe_with_ns.xml")
+
+or
+
+        $ rake console
+        my_xml_string = ""
+        file = File.new("test/nfe_with_ns.xml", "r")
+        while (line = file.gets)
+            my_xml_string = my_xml_string + line
+        end
+        file.close
+
+        xml = RubyDanfe::XML.new(my_xml_string)
+        pdf = RubyDanfe.generatePDF(xml)
+
+        pdf.render_file "output.pdf"
+
+### Building and publishing
+
+You can build using one of the above tasks
+
+        $ rake build    # Build ruby_danfe-X.X.X.gem into the pkg directory
+        $ rake install  # Build and install ruby_danfe-X.X.X.gem into system gems
+        $ rake release  # Create tag vX.X.X and build and push ruby_danfe-X.X.X.gem to Rubygems
+
+## Contributing
+
+We encourage you to contribute to Ruby DANFE!
+
+1. Fork it
+2. Create your feature branch (`git checkout -b my-new-feature`)
+3. Commit your changes (`git commit -am 'Add some feature'`)
+4. Push to the branch (`git push origin my-new-feature`)
+5. Create new Pull Request
+
+## License
+
+Ruby DANFE is released under the [MIT License](http://www.opensource.org/licenses/MIT).
