+++ Rakefile
@@ -1,0 +1,6 @@
+require "bundler/gem_tasks"
+
+desc "Open an irb session preloaded ruby_danfe classes"
+task :console do
+  sh "irb -rubygems -I lib -r ruby_danfe.rb"
+end
