--- build.sh
@@ -1,2 +1,0 @@
-gem build ruby_danfe.gemspec
-gem install ruby_danfe-0.9.4.gem --local --no-ri --no-rdoc
