+++ lib/ruby_danfe/version.rb
@@ -1,0 +1,3 @@
+module RubyDanfe
+  VERSION = "0.9.4"
+end
